+++ src/Utils/SqlStatementsInterface.php
@@ -1,0 +1,44 @@
+<?php
+
+/**
+ * TechDivision\Import\Utils\SqlStatementsInterface
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Utils;
+
+/**
+ * Interfaces of the classes that contains the SQL statements to use.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+interface SqlStatementsInterface
+{
+
+    /**
+     * Returns the SQL statement with the passed name.
+     *
+     * @param string $key The key of the SQL statement to return
+     *
+     * @return string The SQL statement
+     * @throws \Exception Is thrown, if the SQL statement with the passed key cannot be found
+     */
+    public function find($key);
+}